Move input file utils into separate header
