+++ src/core/io/src/4C_io_input_file.cpp
@@ -8,26 +8,15 @@
 #include "4C_io_input_file.hpp"
 
 #include "4C_comm_mpi_utils.hpp"
-#include "4C_fem_general_node.hpp"
-#include "4C_fem_nurbs_discretization_knotvector.hpp"
-#include "4C_io_pstream.hpp"
-#include "4C_linalg_utils_densematrix_communication.hpp"
 #include "4C_utils_string.hpp"
 
 #include <ryml.hpp>
 #include <ryml_std.hpp>
-#include <Teuchos_ParameterList.hpp>
-#include <Teuchos_Time.hpp>
 
 #include <filesystem>
+#include <fstream>
 #include <sstream>
 #include <utility>
-
-#ifdef FOUR_C_ENABLE_FE_TRAPPING
-#include <cfenv>
-#include <format>
-#endif
-
 
 FOUR_C_NAMESPACE_OPEN
 
@@ -35,89 +24,12 @@ namespace Core::IO
 {
   namespace
   {
-    constexpr double tolerance_n = 1.0e-14;
 
     /**
      * Sections that contain at least this number of entries are considered huge and are only
      * available on rank 0.
      */
     constexpr std::size_t huge_section_threshold = 10'000;
-
-    /*----------------------------------------------------------------------*/
-    /*----------------------------------------------------------------------*/
-    Teuchos::ParameterList& find_sublist(std::string name, Teuchos::ParameterList& list)
-    {
-      Teuchos::ParameterList* sublist = &list;
-
-      for (std::string::size_type pos = name.find('/'); pos != std::string::npos;
-          pos = name.find('/'))
-      {
-        sublist = &sublist->sublist(name.substr(0, pos));
-        name = name.substr(pos + 1);
-      }
-
-      return sublist->sublist(name);
-    }
-
-    void add_entry(const std::string& key, const std::string& value, Teuchos::ParameterList& list)
-    {
-      // safety check: Is there a duplicate of the same parameter?
-      if (list.isParameter(key))
-        FOUR_C_THROW("Duplicate parameter '%s' in sublist '%s'", key.c_str(), list.name().c_str());
-
-      if (key.empty()) FOUR_C_THROW("Internal error: missing key.", key.c_str());
-      // safety check: Is the parameter without any specified value?
-      if (value.empty())
-        FOUR_C_THROW("Missing value for parameter %s. Fix your input file!", key.c_str());
-
-      {  // try to find an int
-        std::stringstream ssi;
-        int iv;
-
-        ssi << value;
-        ssi >> iv;
-
-        if (ssi.eof())
-        {
-          list.set(key, iv);
-          return;
-        }
-      }
-
-#ifdef FOUR_C_ENABLE_FE_TRAPPING
-      // somehow the following test whether we have a double or not
-      // creates always an internal floating point exception (FE_INVALID). An alternative
-      // implementation using boost::lexical_cast<double> does not solve this problem!
-      // Better temporarily disable this floating point exception in the following,
-      // so that we can go on.
-      feclearexcept(FE_INVALID);
-      /*feenableexcept(FE_INVALID | FE_DIVBYZERO | FE_UNDERFLOW | FE_OVERFLOW);*/
-      fedisableexcept(FE_INVALID);
-#endif
-
-      {  // try to find a double
-        std::stringstream ssd;
-        double dv;
-
-        ssd << value;
-        ssd >> dv;
-
-#ifdef FOUR_C_ENABLE_FE_TRAPPING
-        feclearexcept(FE_INVALID);
-        /*feenableexcept(FE_INVALID | FE_DIVBYZERO | FE_UNDERFLOW | FE_OVERFLOW);*/
-        feenableexcept(FE_INVALID | FE_DIVBYZERO);
-#endif
-
-        if (ssd.eof())
-        {
-          list.set(key, dv);
-          return;
-        }
-      }
-
-      // if it is not an int or a double it must be a string
-      list.set(key, value);
-    }
 
     //! The different ways we want to handle sections in the input file.
     enum class SectionType
@@ -404,571 +316,6 @@ namespace Core::IO
         content_by_section_.contains(section_name),
         [](const bool& r, const bool& in) { return r || in; }, comm_);
     return known_somewhere;
-  }
-
-
-  /*----------------------------------------------------------------------*/
-  /*----------------------------------------------------------------------*/
-  bool read_parameters_in_section(
-      InputFile& input, const std::string& section_name, Teuchos::ParameterList& list)
-  {
-    if (section_name.empty()) FOUR_C_THROW("Empty section name given.");
-
-    Teuchos::ParameterList& sublist = find_sublist(section_name, list);
-
-    for (const auto& line : input.lines_in_section(section_name))
-    {
-      const auto& [key, value] = read_key_value(std::string(line));
-
-      add_entry(key, value, sublist);
-    }
-
-    return true;
-  }
-
-  /*----------------------------------------------------------------------*/
-  /*----------------------------------------------------------------------*/
-  void read_design(InputFile& input, const std::string& name,
-      std::vector<std::vector<int>>& dobj_fenode,
-      const std::function<const Core::FE::Discretization&(const std::string& name)>&
-          get_discretization)
-  {
-    std::map<int, std::set<int>> topology;
-
-    std::string sectionname = name + "-NODE TOPOLOGY";
-    std::string marker = sectionname;
-
-    // Store lines that need special treatment
-    std::vector<std::string> box_face_conditions;
-
-    for (const auto& l : input.lines_in_section_rank_0_only(marker))
-    {
-      int dobj;
-      int nodeid;
-      std::string nname;
-      std::string dname;
-      std::istringstream stream{std::string(l)};
-      stream >> nname;
-      if (not stream) FOUR_C_THROW("Illegal line in section '%s': '%s'", marker.c_str(), l.data());
-
-      if (nname == "NODE")  // plain old reading of the design nodes from the .dat-file
-      {
-        stream >> nodeid >> dname >> dobj;
-        topology[dobj - 1].insert(nodeid - 1);
-      }
-      else  // fancy specification of the design nodes by specifying min or max of the domain
-      {     // works best on rectangular domains ;)
-        // Store the specification and broadcast it to all processors
-        box_face_conditions.emplace_back(l);
-      }
-    }
-
-    // Broadcast what we have read on rank 0 to all other ranks
-    Core::Communication::broadcast(topology, 0, input.get_comm());
-    Core::Communication::broadcast(box_face_conditions, 0, input.get_comm());
-
-    // Special treatment if we have any box face conditions
-    {
-      for (const auto& l : box_face_conditions)
-      {
-        int dobj;
-        std::string nname;
-        std::string dname;
-        std::string disname;
-        std::array<int, 3> dir = {0, 0, 0};
-
-        std::istringstream stream{l};
-        stream >> nname;
-        if (not stream)
-          FOUR_C_THROW("Illegal line in section '%s': '%s'", marker.c_str(), l.data());
-
-        if (nname == "CORNER" && name == "DNODE")
-        {
-          std::string tmp;
-          stream >> disname;
-          for (int i = 0; i < 3; ++i)
-          {
-            stream >> tmp;
-            if (tmp.size() != 2 || tmp[0] < 'x' || tmp[0] > 'z' || (tmp[1] != '+' && tmp[1] != '-'))
-              FOUR_C_THROW("Illegal design node definition.");
-            dir[tmp[0] - 'x'] = (tmp[1] == '+') ? 1 : -1;
-          }
-          stream >> dname >> dobj;
-        }
-        else if (nname == "EDGE" && name == "DLINE")
-        {
-          std::string tmp;
-          stream >> disname;
-          for (int i = 0; i < 2; ++i)
-          {
-            stream >> tmp;
-            if (tmp.size() != 2 || tmp[0] < 'x' || tmp[0] > 'z' || (tmp[1] != '+' && tmp[1] != '-'))
-              FOUR_C_THROW("Illegal design node definition.");
-            dir[tmp[0] - 'x'] = (tmp[1] == '+') ? 1 : -1;
-          }
-          stream >> dname >> dobj;
-        }
-        else if (nname == "SIDE" && name == "DSURF")
-        {
-          std::string tmp;
-          stream >> disname;
-          stream >> tmp;
-          if (tmp.size() != 2 || tmp[0] < 'x' || tmp[0] > 'z' || (tmp[1] != '+' && tmp[1] != '-'))
-            FOUR_C_THROW("Illegal design node definition.");
-          dir[tmp[0] - 'x'] = (tmp[1] == '+') ? 1 : -1;
-          stream >> dname >> dobj;
-        }
-        else if (nname == "VOLUME" && name == "DVOL")
-        {
-          stream >> disname;
-          stream >> dname >> dobj;
-        }
-        else
-        {
-          FOUR_C_THROW("Illegal line in section '%s': '%s'", marker.c_str(), l.data());
-        }
-
-        const Core::FE::Discretization& actdis = get_discretization(disname);
-
-        std::vector<double> box_specifications;
-        {
-          for (int init = 0; init < 9; ++init) box_specifications.push_back(0.0);
-          if (Core::Communication::my_mpi_rank(input.get_comm()) == 0)  // Reading is done by proc 0
-          {
-            // get original domain section from the *.dat-file
-            std::string dommarker = disname + " DOMAIN";
-            std::transform(dommarker.begin(), dommarker.end(), dommarker.begin(), ::toupper);
-
-            for (const auto& line : input.lines_in_section_rank_0_only(dommarker))
-            {
-              std::istringstream t{std::string{line}};
-              std::string key;
-              t >> key;
-
-              if (key == "LOWER_BOUND")
-              {
-                t >> box_specifications[0] >> box_specifications[1] >> box_specifications[2];
-              }
-              else if (key == "UPPER_BOUND")
-              {
-                t >> box_specifications[3] >> box_specifications[4] >> box_specifications[5];
-              }
-              else if (key == "ROTATION")
-              {
-                t >> box_specifications[6] >> box_specifications[7] >> box_specifications[8];
-              }
-            }
-          }
-          // All other processors get this info broadcasted
-          Core::Communication::broadcast(box_specifications.data(),
-              static_cast<int>(box_specifications.size()), 0, input.get_comm());
-        }
-
-        // determine the active discretizations bounding box
-        std::array<double, 6> bbox;
-        for (size_t i = 0; i < sizeof(bbox) / sizeof(bbox[0]); ++i) bbox[i] = box_specifications[i];
-
-        // manipulate the bounding box according to the specified condition
-        for (size_t i = 0; i < 3; ++i)
-        {
-          switch (dir[i])
-          {
-            case 0:
-              bbox[i + 0] = std::numeric_limits<double>::max();
-              bbox[i + 3] = -std::numeric_limits<double>::max();
-              break;
-            case -1:
-              bbox[i] += tolerance_n;
-              bbox[i + 3] = std::numeric_limits<double>::max();
-              break;
-            case 1:
-              bbox[i] = -std::numeric_limits<double>::max();
-              bbox[i + 3] -= tolerance_n;
-              break;
-            default:
-              FOUR_C_THROW("Invalid BC specification");
-          }
-        }
-
-        // collect all nodes which are outside the adapted bounding box
-        std::set<int> dnodes;
-        for (const auto* node : actdis.my_row_node_range())
-        {
-          const auto& coord = node->x();
-          std::array<double, 3> coords;
-          coords[0] = coord[0];
-          coords[1] = coord[1];
-          coords[2] = coord[2];
-          // rotate back to identify condition, if a rotation is defined
-          static const int rotoffset = 6;
-          for (int rotaxis = 2; rotaxis > -1; --rotaxis)
-          {
-            if (box_specifications[rotaxis + rotoffset] != 0.0)
-            {
-              std::array<double, 3> coordm;
-              coordm[0] = (box_specifications[0] + box_specifications[3]) / 2.;
-              coordm[1] = (box_specifications[1] + box_specifications[4]) / 2.;
-              coordm[2] = (box_specifications[2] + box_specifications[5]) / 2.;
-              // add rotation around mitpoint here.
-              std::array<double, 3> dx;
-              dx[0] = coords[0] - coordm[0];
-              dx[1] = coords[1] - coordm[1];
-              dx[2] = coords[2] - coordm[2];
-
-              double calpha = cos(-box_specifications[rotaxis + rotoffset] * M_PI / 180);
-              double salpha = sin(-box_specifications[rotaxis + rotoffset] * M_PI / 180);
-
-              coords[0] = coordm[0];  //+ calpha*dx[0] + salpha*dx[1];
-              coords[1] = coordm[1];  //+ -salpha*dx[0] + calpha*dx[1];
-              coords[2] = coordm[2];
-
-              coords[(rotaxis + 1) % 3] +=
-                  calpha * dx[(rotaxis + 1) % 3] + salpha * dx[(rotaxis + 2) % 3];
-              coords[(rotaxis + 2) % 3] +=
-                  calpha * dx[(rotaxis + 2) % 3] - salpha * dx[(rotaxis + 1) % 3];
-              coords[rotaxis] += dx[rotaxis];
-            }
-          }
-
-          if ((coords[0] <= bbox[0] || coords[0] >= bbox[3]) &&
-              (coords[1] <= bbox[1] || coords[1] >= bbox[4]) &&
-              (coords[2] <= bbox[2] || coords[2] >= bbox[5]))
-            dnodes.insert(node->id());
-        }
-        Core::LinAlg::gather_all(dnodes, input.get_comm());
-        topology[dobj - 1].insert(dnodes.begin(), dnodes.end());
-
-
-        if (dname.substr(0, name.length()) != name)
-          FOUR_C_THROW("Illegal line in section '%s': '%s'\n%s found, where %s was expected",
-              marker.c_str(), l.data(), dname.substr(0, name.length()).c_str(), name.c_str());
-      }
-    }
-
-    if (topology.size() > 0)
-    {
-      int max_num_dobj = topology.rbegin()->first;
-      if (max_num_dobj >= static_cast<int>(dobj_fenode.size()))
-        dobj_fenode.resize(max_num_dobj + 1);
-      // copy all design object entries
-      for (auto& topo : topology)
-      {
-        // we copy from a std::set, thus the gids are sorted
-        dobj_fenode[topo.first].reserve(topo.second.size());
-        dobj_fenode[topo.first].assign(topo.second.begin(), topo.second.end());
-      }
-    }
-  }
-
-
-  //----------------------------------------------------------------------
-  /// read a knotvector section (for isogeometric analysis)
-  //----------------------------------------------------------------------
-  void read_knots(InputFile& input, const std::string& name,
-      std::shared_ptr<Core::FE::Nurbs::Knotvector>& disknots)
-  {
-    // io to shell
-    const int myrank = Core::Communication::my_mpi_rank(input.get_comm());
-
-    Teuchos::Time time("", true);
-
-    // only the knotvector section of this discretisation
-    // type is of interest
-    std::string field;
-    if (name == "fluid" or name == "xfluid" or name == "porofluid")
-    {
-      field = "FLUID";
-    }
-    else if (name == "structure")
-    {
-      field = "STRUCTURE";
-    }
-    else if (name == "ale")
-    {
-      field = "ALE";
-    }
-    else if (name == "scatra")
-    {
-      field = "TRANSPORT";
-    }
-    else if (name == "thermo")
-    {
-      field = "THERMO";
-    }
-    else if (name == "scatra_micro")
-    {
-      field = "TRANSPORT2";
-    }
-    else
-    {
-      FOUR_C_THROW("Unknown discretization name for knotvector input\n");
-    }
-
-    // another valid section name was found
-    const std::string sectionname = field + " KNOTVECTORS";
-
-    if (myrank == 0)
-    {
-      Core::IO::cout << "Reading knot vectors for " << name << " discretization :\n";
-      fflush(stdout);
-    }
-
-    // number of patches to be determined
-    int npatches = 0;
-
-    // dimension of nurbs patches
-    int nurbs_dim = 0;
-
-    //--------------------------------------------------------------------
-    //--------------------------------------------------------------------
-    //      first, determine number of patches and dimension of nurbs
-    //--------------------------------------------------------------------
-    //--------------------------------------------------------------------
-    {
-      // temporary string
-      std::string tmp;
-      // loop lines in file
-      for (const auto& line : input.lines_in_section(sectionname))
-      {
-        // count number of patches in knotvector section of
-        // this discretisation
-        {
-          std::string::size_type loc;
-          std::istringstream file{std::string{line}};
-          file >> tmp;
-
-          // check for the number of dimensions
-          loc = tmp.rfind("NURBS_DIMENSION");
-          if (loc != std::string::npos)
-          {
-            // set number of nurbs dimension
-            std::string str_nurbs_dim;
-            file >> str_nurbs_dim;
-            char* endptr = nullptr;
-            nurbs_dim = static_cast<int>(strtol(str_nurbs_dim.c_str(), &endptr, 10));
-
-            continue;
-          }
-
-          // check for a new patch
-          loc = tmp.rfind("ID");
-          if (loc != std::string::npos)
-          {
-            // increase number of patches
-            npatches++;
-
-            continue;
-          }
-        }
-      }  // end loop through file
-    }
-
-    if (myrank == 0)
-    {
-      printf("                        %8d patches", npatches);
-      fflush(stdout);
-    }
-
-
-    //--------------------------------------------------------------------
-    //--------------------------------------------------------------------
-    //                alloc knotvector object to fill
-    //--------------------------------------------------------------------
-    //--------------------------------------------------------------------
-
-    // allocate knotvector for this dis
-    disknots = std::make_shared<Core::FE::Nurbs::Knotvector>(nurbs_dim, npatches);
-
-    // make sure that we have some Knotvector object to fill
-    if (disknots == nullptr)
-    {
-      FOUR_C_THROW("disknots should have been allocated before");
-    }
-
-    //--------------------------------------------------------------------
-    //--------------------------------------------------------------------
-    //                finally read knotvector section
-    //--------------------------------------------------------------------
-    //--------------------------------------------------------------------
-    {
-      // this is a pointer to the knots of one patch in one direction
-      // we will read them and put them
-      std::vector<std::shared_ptr<std::vector<double>>> patch_knots(nurbs_dim);
-
-      // temporary string
-      std::string tmp;
-
-      // start to read something when read is true
-      bool read = false;
-
-      // index for number of patch
-      int npatch = 0;
-      // index for u/v/w
-      int actdim = -1;
-      // ints for the number of knots
-      std::vector<int> n_x_m_x_l(nurbs_dim);
-      // ints for patches degrees
-      std::vector<int> degree(nurbs_dim);
-      // a vector of strings holding the knotvectortypes read
-      std::vector<std::string> knotvectortype(nurbs_dim);
-
-      // count for sanity check
-      int count_read = 0;
-      std::vector<int> count_vals(nurbs_dim);
-
-      // loop lines in file
-      for (const auto& line : input.lines_in_section(sectionname))
-      {
-        std::istringstream file{std::string{line}};
-        file >> tmp;
-
-        // check for a new patch
-        std::string::size_type loc = tmp.rfind("BEGIN");
-        if (loc != std::string::npos)
-        {
-          file >> tmp;
-
-          // activate reading
-          read = true;
-
-          actdim = -1;
-
-          // create vectors for knots in this patch
-          for (int rr = 0; rr < nurbs_dim; ++rr)
-          {
-            patch_knots[rr] = std::make_shared<std::vector<double>>();
-            (*(patch_knots[rr])).clear();
-          }
-
-          // reset counter for knot values
-          for (int rr = 0; rr < nurbs_dim; rr++)
-          {
-            count_vals[rr] = 0;
-          }
-
-          continue;
-        }
-
-        // get ID of patch we are currently reading
-        loc = tmp.rfind("ID");
-        if (loc != std::string::npos)
-        {
-          std::string str_npatch;
-          file >> str_npatch;
-
-          char* endptr = nullptr;
-          npatch = static_cast<int>(strtol(str_npatch.c_str(), &endptr, 10));
-          npatch--;
-
-          continue;
-        }
-
-        // get number of knots in the knotvector direction
-        // we are currently reading
-        loc = tmp.rfind("NUMKNOTS");
-        if (loc != std::string::npos)
-        {
-          std::string str_numknots;
-          file >> str_numknots;
-
-          // increase dimesion for knotvector (i.e. next time
-          // we'll fill the following knot vector)
-          actdim++;
-          if (actdim > nurbs_dim)
-          {
-            FOUR_C_THROW(
-                "too many knotvectors, we only need one for each dimension (nurbs_dim = %d)\n",
-                nurbs_dim);
-          }
-
-          char* endptr = nullptr;
-          n_x_m_x_l[actdim] = static_cast<int>(strtol(str_numknots.c_str(), &endptr, 10));
-
-          continue;
-        }
-
-        // get number of bspline polinomial associated with
-        // knots in this direction
-        loc = tmp.rfind("DEGREE");
-        if (loc != std::string::npos)
-        {
-          std::string str_degree;
-          file >> str_degree;
-
-          char* endptr = nullptr;
-          degree[actdim] = static_cast<int>(strtol(str_degree.c_str(), &endptr, 10));
-
-          continue;
-        }
-
-        // get type of knotvector (interpolated or periodic)
-        loc = tmp.rfind("TYPE");
-        if (loc != std::string::npos)
-        {
-          std::string type;
-
-          file >> type;
-          knotvectortype[actdim] = type;
-
-          continue;
-        }
-
-        // locate end of patch
-        loc = tmp.rfind("END");
-        if (loc != std::string::npos)
-        {
-          for (int rr = 0; rr < nurbs_dim; ++rr)
-          {
-            disknots->set_knots(
-                rr, npatch, degree[rr], n_x_m_x_l[rr], knotvectortype[rr], patch_knots[rr]);
-          }
-          file >> tmp;
-          // stop reading of knot values if we are here
-          read = false;
-
-          for (int rr = 0; rr < nurbs_dim; rr++)
-          {
-            if (n_x_m_x_l[rr] != count_vals[rr])
-            {
-              FOUR_C_THROW("not enough knots read in dim %d (%d!=NUMKNOTS=%d), nurbs_dim=%d\n", rr,
-                  count_vals[rr], n_x_m_x_l[rr], nurbs_dim);
-            }
-          }
-
-          // count for sanity check
-          count_read++;
-
-          continue;
-        }
-
-        //  reading of knot values if read is true and no
-        // other keyword was found
-        if (read)
-        {
-          char* endptr = nullptr;
-
-          double dv = strtod(tmp.c_str(), &endptr);
-
-          // count for sanity check
-          count_vals[actdim]++;
-
-          (*(patch_knots[actdim])).push_back(dv);
-        }
-      }  // end loop through file
-
-      if (count_read != npatches)
-      {
-        FOUR_C_THROW("wasn't able to read enough patches\n");
-      }
-    }
-
-    if (myrank == 0)
-    {
-      Core::IO::cout << " in...." << time.totalElapsedTime(true) << " secs\n";
-
-      time.reset();
-      fflush(stdout);
-    }
   }
