+++ src/core/io/tests/4C_io_input_file_test.cpp
@@ -9,6 +9,7 @@
 
 #include "4C_io_input_file.hpp"
 
+#include "4C_io_input_file_utils.hpp"
 #include "4C_unittest_utils_assertions_test.hpp"
 #include "4C_unittest_utils_support_files_test.hpp"
 #include "4C_utils_exceptions.hpp"
@@ -19,60 +20,7 @@ namespace
 {
   using namespace FourC;
 
-  TEST(ReadKeyValue, WithWhitespace)
-  {
-    const auto& [key, value] = Core::IO::read_key_value("key 1.0");
-    EXPECT_EQ(key, "key");
-    EXPECT_EQ(value, "1.0");
-  }
 
-  TEST(ReadKeyValue, WithWhitespaceMultipleTakesFirst)
-  {
-    const auto& [key, value] = Core::IO::read_key_value("key 1.0 2.0 3");
-    EXPECT_EQ(key, "key");
-    EXPECT_EQ(value, "1.0 2.0 3");
-  }
-
-  TEST(ReadKeyValue, WithWhitespaceAndEqualSignInside)
-  {
-    const auto& [key, value] = Core::IO::read_key_value("key=key value=value");
-    EXPECT_EQ(key, "key=key");
-    EXPECT_EQ(value, "value=value");
-  }
-
-  TEST(ReadKeyValue, WithEqualsSign)
-  {
-    const auto& [key, value] = Core::IO::read_key_value("key = 1.0");
-    EXPECT_EQ(key, "key");
-    EXPECT_EQ(value, "1.0");
-  }
-
-  TEST(ReadKeyValue, WithEqualsSignMultipleTakesFirst)
-  {
-    const auto& [key, value] = Core::IO::read_key_value("key = 1.0 = 2.0 = 3.0=4.0");
-    EXPECT_EQ(key, "key");
-    EXPECT_EQ(value, "1.0 = 2.0 = 3.0=4.0");
-  }
-
-  TEST(ReadKeyValue, WithEqualsSignNoKey)
-  {
-    EXPECT_ANY_THROW(Core::IO::read_key_value("   = 1.0"));
-  }
-
-  TEST(ReadKeyValue, WithEqualsSignNoKValue)
-  {
-    EXPECT_ANY_THROW(Core::IO::read_key_value(" key   =      "));
-  }
-
-  TEST(ReadKeyValue, SingleWordThrows) { EXPECT_ANY_THROW(Core::IO::read_key_value("key")); }
-
-
-  TEST(ReadKeyValue, EmptyThrows) { EXPECT_ANY_THROW(Core::IO::read_key_value("")); };
-
-  TEST(ReadKeyValue, WithEqualsSignNoSpaceThrows)
-  {
-    EXPECT_ANY_THROW(Core::IO::read_key_value("key=1.0"));
-  }
 
   void check_section(
       Core::IO::InputFile& input, const std::string& section, const std::vector<std::string>& lines)
