+++ src/core/io/tests/4C_io_input_file_utils_test.cpp
@@ -19,6 +19,61 @@ namespace
 {
   using namespace FourC;
 
+  TEST(ReadKeyValue, WithWhitespace)
+  {
+    const auto& [key, value] = Core::IO::read_key_value("key 1.0");
+    EXPECT_EQ(key, "key");
+    EXPECT_EQ(value, "1.0");
+  }
+
+  TEST(ReadKeyValue, WithWhitespaceMultipleTakesFirst)
+  {
+    const auto& [key, value] = Core::IO::read_key_value("key 1.0 2.0 3");
+    EXPECT_EQ(key, "key");
+    EXPECT_EQ(value, "1.0 2.0 3");
+  }
+
+  TEST(ReadKeyValue, WithWhitespaceAndEqualSignInside)
+  {
+    const auto& [key, value] = Core::IO::read_key_value("key=key value=value");
+    EXPECT_EQ(key, "key=key");
+    EXPECT_EQ(value, "value=value");
+  }
+
+  TEST(ReadKeyValue, WithEqualsSign)
+  {
+    const auto& [key, value] = Core::IO::read_key_value("key = 1.0");
+    EXPECT_EQ(key, "key");
+    EXPECT_EQ(value, "1.0");
+  }
+
+  TEST(ReadKeyValue, WithEqualsSignMultipleTakesFirst)
+  {
+    const auto& [key, value] = Core::IO::read_key_value("key = 1.0 = 2.0 = 3.0=4.0");
+    EXPECT_EQ(key, "key");
+    EXPECT_EQ(value, "1.0 = 2.0 = 3.0=4.0");
+  }
+
+  TEST(ReadKeyValue, WithEqualsSignNoKey)
+  {
+    EXPECT_ANY_THROW(Core::IO::read_key_value("   = 1.0"));
+  }
+
+  TEST(ReadKeyValue, WithEqualsSignNoKValue)
+  {
+    EXPECT_ANY_THROW(Core::IO::read_key_value(" key   =      "));
+  }
+
+  TEST(ReadKeyValue, SingleWordThrows) { EXPECT_ANY_THROW(Core::IO::read_key_value("key")); }
+
+
+  TEST(ReadKeyValue, EmptyThrows) { EXPECT_ANY_THROW(Core::IO::read_key_value("")); };
+
+  TEST(ReadKeyValue, WithEqualsSignNoSpaceThrows)
+  {
+    EXPECT_ANY_THROW(Core::IO::read_key_value("key=1.0"));
+  }
+
   TEST(YamlMetadata, Dump)
   {
     Teuchos::ParameterList pl;
@@ -33,7 +88,7 @@ namespace
     pl_AB.set("f", "string2");
 
     std::stringstream out;
-    Core::IO::InputFileUtils::print_metadata_yaml(out, pl);
+    Core::IO::print_metadata_yaml(out, pl);
 
     const std::string expected_output =
         R"(parameters:
